NR NGAP: Fix Time UE Stayed in Cell IE type and calculation
According to 3GPP TS 38.413, the 'Time UE Stayed in Cell' IE is an
integer ranging from 0 to 4095. This patch updates the data structure
to reflect the correct bit width and implements the dynamic calculation
based on the UE's residence time.

- Change time_in_cell from long to uint16_t in ngap_messages_types.h.
- Replace dummy value with real-time calculation in rrc_gNB_NGAP.c.
- Implement a ceiling of 4095 to comply with the 12-bit protocol limit.

Ref: 3GPP TS 38.413 Section 9.3.1.97

+++ openair2/COMMON/ngap_messages_types.h
@@ -474,7 +474,7 @@ typedef struct {
   // Cell type
   uint8_t type;
   // Time UE Stayed in Cell
-  long time_in_cell;
+  uint16_t time_in_cell;
   // Cause
   ngap_cause_t *cause;
 } last_visited_ngran_cell_info_t;
